refactor change id uint type to uint64

+++ internal/project/interfaces.go
@@ -9,12 +9,12 @@ import (
 type Writer interface {
 	Insert(ctx context.Context, project *entity.Project) error
 	Update(ctx context.Context, project *entity.Project) error
-	Delete(ctx context.Context, id uint) error
+	Delete(ctx context.Context, id uint64) error
 }
 
 type Reader interface {
 	FindAll(ctx context.Context) ([]*entity.Project, error)
-	FindByID(ctx context.Context, id uint) (*entity.Project, error)
+	FindByID(ctx context.Context, id uint64) (*entity.Project, error)
 }
 
 type Repository interface {
@@ -24,8 +24,8 @@ type Repository interface {
 
 type Service interface {
 	Create(ctx context.Context, project *entity.Project) error
-	GetByID(ctx context.Context, id uint) (*entity.Project, error)
+	GetByID(ctx context.Context, id uint64) (*entity.Project, error)
 	GetAll(ctx context.Context) ([]*entity.Project, error)
 	Update(ctx context.Context, project *entity.Project) error
-	Delete(ctx context.Context, id uint) error
+	Delete(ctx context.Context, id uint64) error
 }
